+++ CHANGELOG.rst
@@ -30,6 +30,7 @@ Changed
     - Reduce Django queries to optimize iRODS file search (#393)
     - Replace IRODS query limit settings with ``SHEETS_IRODS_LIMIT`` (#393)
     - Cancer study app: only show shortcuts for genome/exome seq assays (#398)
+    - Move germline specific template tags in germline study app (#399)
 
 Fixed
 -----